Reconfigured to use build-key-pass instead of build-key
The "build-key" script creates a client configuration that has the key unencrypted in the file. This would allow anyone with the file to access the VPN with no further checks. I've converted this to use the "build-key-pass" script which prompts the user for a password to connect to the VPN.